requirements.txt
Flask==2.2.3
flask-socketio==5.2.0
pandas==1.5.3
transformers==4.28.0
matplotlib==3.6.2
spacy==3.4.3
textblob==0.15.3
langdetect==1.0.9
deep-translator==1.8.0
requests==2.28.2
python-dotenv==0.21.0
google-generativeai==0.3.0
eventlet==0.33.0
gunicorn==20.1.0
werkzeug==2.2.3
numpy==1.23.5
eventlet==0.33.0
dnspython==2.3.0
gevent